+++ src/init.c
@@ -22,6 +22,7 @@ extern SEXP vctrs_equal_object(SEXP, SEXP);
 extern SEXP vctrs_in(SEXP, SEXP);
 extern SEXP vctrs_duplicated(SEXP);
 extern SEXP vctrs_unique_loc(SEXP);
+extern SEXP vctrs_duplicate_loc(SEXP);
 extern SEXP vctrs_count(SEXP);
 extern SEXP vctrs_id(SEXP);
 extern SEXP vctrs_n_distinct(SEXP);
@@ -121,6 +122,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_equal_object",               (DL_FUNC) &vctrs_equal_object, 2},
   {"vctrs_in",                         (DL_FUNC) &vctrs_in, 2},
   {"vctrs_unique_loc",                 (DL_FUNC) &vctrs_unique_loc, 1},
+  {"vctrs_duplicate_loc",              (DL_FUNC) &vctrs_duplicate_loc, 1},
   {"vctrs_duplicated",                 (DL_FUNC) &vctrs_duplicated, 1},
   {"vctrs_duplicated_any",             (DL_FUNC) &vctrs_duplicated_any, 1},
   {"vctrs_count",                      (DL_FUNC) &vctrs_count, 1},